Move auth into a separate module
Move auth and login related functionality into a separate auth module.

To clean up the code, the auth state is pulled into an auth context
which is provided to the relevant components through the AuthProvider
helper wrapper. This allows the global auth state to be easily shared
amongst any component that needs it.