Re-enable Google login button
The one-tap feature has an automatic back-off if a user clicks out of
it, preventing the user from logging into the site at all. The log-in
button is required to allow for this re-login

Additionally, this requires refactoring of the logic to ensure the
google sdk is initialized before being used elsewhere. Like the auth
state this is now handled via a context.